backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg[binary]==3.1.18
psycopg-pool==3.2.1
requests==2.31.0
google-auth==2.29.0
google-api-python-client==2.126.0
PyJWT==2.8.0
